+++ src/test/java/org/osgl/MappingTest.java
@@ -22,6 +22,8 @@ package org.osgl;
 
 import static org.osgl.Lang.requireNotNull;
 
+import com.alibaba.fastjson.JSON;
+import com.alibaba.fastjson.JSONObject;
 import org.joda.time.DateTime;
 import org.junit.*;
 import org.junit.experimental.runners.Enclosed;